requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.07.19=hca03da5_0
certifi=2022.9.24=py310hca03da5_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.2=hc377ac9_4
ncurses=6.3=h1a28f6b_3
numpy=1.23.4=pypi_0
openssl=1.1.1q=h1a28f6b_0
pip=22.2.2=py310hca03da5_0
pyserial=3.5=pypi_0
python=3.10.6=hbdb9e5c_0
readline=8.1.2=h1a28f6b_1
setuptools=63.4.1=py310hca03da5_0
sqlite=3.39.3=h1058600_0
tk=8.6.12=hb8d0fd4_0
tzdata=2022e=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h1a28f6b_0
zlib=1.2.12=h5a0b063_3
